% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proliferation_modeling_functions.r
\name{find.local.max}
\alias{find.local.max}
\title{Find Local Max}
\usage{
find.local.max(x, y, lower.x, upper.x)
}
\arguments{
\item{x}{vector of numbers matching y in length.}

\item{y}{vector of numbers matching x in length.}

\item{lower.x}{lower limit of x to search.}

\item{upper.x}{upper limit of x to search.}
}
\value{
the value of x where y is maximised.
}
\description{
Find the maximum value of y within a range of x values
}
\examples{
x <- rnorm(100)
y <- rnorm(100)
find.local.max(x, y, -1, 1)
}
