% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proliferation_modeling_functions.r
\name{find.next.peak}
\alias{find.next.peak}
\title{Find the mode of the next peak using the previous peak and the average peak
distance}
\usage{
find.next.peak(x, y, prev.peak, peak.dist)
}
\description{
Find the mode of the next peak using the previous peak and the average peak
distance
}
