% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proliferation_modeling_functions.r
\name{find.peak.x.approx.mode}
\alias{find.peak.x.approx.mode}
\title{Find the mode of the first left sided peak in the data, given some
constraints.}
\usage{
find.peak.x.approx.mode(
  trace,
  peak.x.thresh.summit,
  peak.x.thresh.enrich,
  peak.x.upper.bound = NULL,
  ...
)
}
\description{
Find the mode of the first left sided peak in the data, given some
constraints.
}
