% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_detection.R
\name{find_enrichment}
\alias{find_enrichment}
\title{Find the relative enrichment over a valley}
\usage{
find_enrichment(x, y, curpeak.mean, est.window)
}
\description{
Find the relative enrichment over a valley
}
