% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_detection.R
\name{find_initial_peaks}
\alias{find_initial_peaks}
\title{Estimate initial parameters for proliferation model}
\usage{
find_initial_peaks(
  x,
  y,
  peak.0.lower.bound,
  peak.thresh.enrich = 1,
  peak.thresh.summit = 0.05,
  peak.max = 12,
  plot = F,
  window.scaling.factors = c(0.25, 0.25)
)
}
\arguments{
\item{x}{x values (histogram midpoints)}

\item{y}{a smoothed or raw count trace}

\item{peak.0.lower.bound}{lower bound for finding gen0}

\item{peak.thresh.enrich}{fold change over valley to call peak in initial estimation (default 1)}

\item{peak.thresh.summit}{minimum height of a peak in percentage of total heights (default 0.05)}

\item{peak.max}{the maximum number of peaks to search for (default 12)}

\item{plot}{should plot be generated}

\item{window.scaling.factors}{Scaling factors for peak differences. See details (default c(0.25, 0.25))}
}
\value{
data frame with initial peak estimates
}
\description{
Estimate initial parameters for proliferation model
}
\details{
`window.scaling.factors`
window.scaling.factors control the size of the window left and right around the next peak position
to find the mode of the next peak. The smaller this value is, the closer to the half intensity the
peak estimates will be.

I.e. if the estimated distance between peaks is 0.5 and the scaling
factors are c(0.25, 0.5) and the current peak is 1.5 the next peak mode will be estimates as:
lower <- 0.5 * 0.25
upper <- 0.5 * 0.5

# Where should the next peak be based on half it's intensity
est.peak.pos <- log10(10^1.5/2)

Then find the max value in the window est.peak.pos-lower, est.peak.pos+upper
}
