% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/peak_detection.R
\name{find.peak.x.approx.mode}
\alias{find.peak.x.approx.mode}
\alias{find_peak_x_approx_mode}
\title{Find the mode of the first left sided peak in the data, given some constraints.}
\usage{
find.peak.x.approx.mode(...)

find_peak_x_approx_mode(
  trace,
  peak.x.thresh.summit,
  peak.x.thresh.enrich,
  peak.x.upper.bound = NULL,
  ...
)
}
\arguments{
\item{trace}{raw FACS intensities (log10 scale)}

\item{peak.x.thresh.summit}{minimum relative height of peak x}

\item{peak.x.thresh.enrich}{minimum enrichment of peak x over valley}

\item{peak.x.upper.bound}{hard upper limit for peak x position}
}
\description{
Find the mode of the first left sided peak in the data, given some constraints.
}
