% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_cocult_functions.r
\name{gaussian.prolif.model}
\alias{gaussian.prolif.model}
\title{Gaussian prolif model, it is a mixture distribution scaled by relative
peak heights.}
\usage{
gaussian.prolif.model(means, sd, summits, x)
}
\description{
Gaussian prolif model, it is a mixture distribution scaled by relative
peak heights.
}
