% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_cocult_functions.r
\name{gaussian.prolif.model.wrapper}
\alias{gaussian.prolif.model.wrapper}
\title{Wrapper that returns a gaussian proliferation model.
Takes a vector of paremeters par, number of peaks, and data X}
\usage{
gaussian.prolif.model.wrapper(par, n.peaks, x, fixed = NULL)
}
\description{
Wrapper that returns a gaussian proliferation model.
Takes a vector of paremeters par, number of peaks, and data X
}
