% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_cocult_functions.r
\name{gaussian.prolif.resid}
\alias{gaussian.prolif.resid}
\title{Residuals between observed CTV trace and gaussian prolif model}
\usage{
gaussian.prolif.resid(par, n.peaks, x, y, fixed = NULL)
}
\description{
Residuals between observed CTV trace and gaussian prolif model
}
