% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_cocult_functions.r
\name{gaussian.prolif.single.peak}
\alias{gaussian.prolif.single.peak}
\title{Gaussian of a single peak}
\usage{
gaussian.prolif.single.peak(x, mean, sd, summit)
}
\description{
Gaussian of a single peak
}
