% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_cocult_functions.r
\name{get.prolif.stats}
\alias{get.prolif.stats}
\title{Get proliferation statistics}
\usage{
get.prolif.stats(peak.stats)
}
\arguments{
\item{peak.stats}{output from fit.peaks (data frame)}
}
\value{
A dataframe with proliferation statistics
}
\description{
Takes output from fit.peaks and calculates proliferation statistics.
Proliferation statistics according to the FCS express definitions
https://fcsexpressdownloads.s3.amazonaws.com/manual/manual_WIN_RUO/index.html?proliferation_statistics.htm
}
\details{
Can be used manually by provided a data frame with two columns

- generation

- peak_events

Where each row represent the peak number and number of events in a peak.
}
\examples{

# Simulate proliferation data
y <- 10 ^ rnorm(1000, mean=10, sd=0.05)
y <- c(y/4, y/2, y)

# Fit peaks
peaks <- fit.peaks.ls(y, 0, 0, peak.0.lower.bound=9.8)
get.prolif.stats(peaks)

# Manual table
peaks <- data.frame(generation=c(0,1,2), peak_events=c(1000, 1000, 1000))
get.prolif.stats(peaks)
}
