% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/fitting.R
\name{get.prolif.stats}
\alias{get.prolif.stats}
\alias{get_prolif_stats}
\title{Get proliferation statistics}
\usage{
get.prolif.stats(...)

get_prolif_stats(peak.stats)
}
\arguments{
\item{peak.stats}{output from fit_peaks (data frame)}
}
\value{
A dataframe with proliferation statistics
}
\description{
Takes output from fit_peaks and calculates proliferation statistics.
Proliferation statistics according to the FCS express definitions
https://fcsexpressdownloads.s3.amazonaws.com/manual/manual_WIN_RUO/index.html?proliferation_statistics.htm
}
\details{
Can be used manually by provided a data frame with two columns

- generation

- peak_events

Where each row represent the peak number and number of events in a peak.
}
\examples{

# Simulate proliferation data
y <- 10 ^ rnorm(1000, mean=10, sd=0.05)
y <- c(y/4, y/2, y)

# Fit peaks
peaks <- fit_peaks(y, peak.0.lower.bound=9.8)
get_prolif_stats(peaks)

# Manual table
peaks <- data.frame(generation=c(0,1,2), peak_events=c(1000, 1000, 1000))
get_prolif_stats(peaks)
}
