% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proliferation_modeling_functions.r
\name{nearest.index}
\alias{nearest.index}
\title{Find nearest index}
\usage{
nearest.index(x, value)
}
\arguments{
\item{x}{a vector}

\item{the}{value to search the nearest value in x}
}
\value{
the index in x closest to value
}
\description{
Find the index of the value in x that is closest to value
}
\examples{
x <- rnorm(100)
nearest.index(x, 0)
}
