% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nearest_index}
\alias{nearest_index}
\title{Find nearest index}
\usage{
nearest_index(x, value)
}
\arguments{
\item{x}{a vector}

\item{value}{the value to search the nearest value in x}
}
\value{
the index in x closest to value
}
\description{
Find the index of the value in x that is closest to value
}
\examples{
x <- rnorm(100)
nearest_index(x, 0)
}
