% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proliferation_modeling_functions.r
\name{nn.smoother}
\alias{nn.smoother}
\title{Nearest Neighbour smoother}
\usage{
nn.smoother(y, window = 2)
}
\arguments{
\item{y}{a vector y}

\item{window}{how many values up and down are averaged}
}
\value{
the mean of each y
value and the surrounding values specified by window
}
\description{
Smooth a trace using a kernel of size window.
}
\examples{
y <- rnorm(100)
nn.smoother(y, 2)
}
