% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proliferation_modeling_functions.r
\name{opt.new.env}
\alias{opt.new.env}
\title{Create a new env for storing optimization trace}
\usage{
opt.new.env(n.peaks)
}
\description{
Create a new env for storing optimization trace
}
