% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proliferation_plotting_functions.r
\name{opt.plot.estimates}
\alias{opt.plot.estimates}
\title{Plot intial parameter estimates}
\usage{
opt.plot.estimates(
  cur.hist,
  y.smth,
  peak.stats,
  peak.0.lower.bound,
  peak.x.model,
  peak.x.mode,
  peak.x.upper.bound
)
}
\description{
Plot intial parameter estimates
}
