% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proliferation_plotting_functions.r
\name{opt.plot.final.pp}
\alias{opt.plot.final.pp}
\title{Plot optimization of parameters per peak}
\usage{
opt.plot.final.pp(x.mids, y, ps, main = "Optimal fit vs data")
}
\description{
Plot optimization of parameters per peak
}
