% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_cocult_functions.r
\name{opt.prolif.model.ls}
\alias{opt.prolif.model.ls}
\title{Optimize a proliferation model using least squares}
\usage{
opt.prolif.model.ls(
  x,
  y,
  starts,
  upper,
  lower,
  fixed,
  peak.stats,
  plot = F,
  opt.env = NULL
)
}
\description{
Optimize a proliferation model using least squares
}
