% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{opt_plot_final_pp}
\alias{opt_plot_final_pp}
\title{Plot optimization of parameters per peak}
\usage{
opt_plot_final_pp(x.mids, y, ps, main = "Optimal fit vs data")
}
\description{
Plot optimization of parameters per peak
}
