% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{opt_plot_params}
\alias{opt_plot_params}
\title{Plot optimization trace showing convergence of parameters}
\usage{
opt_plot_params(x, opt.env)
}
\description{
Plot optimization trace showing convergence of parameters
}
