% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_cocult_functions.r
\name{prolif.model.density}
\alias{prolif.model.density}
\title{Density of mixture distribution of Gaussian}
\usage{
prolif.model.density(means, sd, summits, x, verbose = F)
}
\description{
Density of mixture distribution of Gaussian
}
