% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_cocult_functions.r
\name{prolif.model.nll}
\alias{prolif.model.nll}
\title{Negative log likelihood of mixture distribution of Gaussian}
\usage{
prolif.model.nll(means, sd, summits, x, verbose = F)
}
\description{
Negative log likelihood of mixture distribution of Gaussian
}
