% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_cocult_functions.r
\name{prolif.resid}
\alias{prolif.resid}
\title{Residuals between observed CTV trace and Gaussian prolif model}
\usage{
prolif.resid(par, n.peaks, x, y, fixed = NULL, opt.env = NULL)
}
\description{
Residuals between observed CTV trace and Gaussian prolif model
}
