% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proliferation_modeling_functions.r
\name{prolif.single.peak}
\alias{prolif.single.peak}
\title{Scaled distribution of a single peak in prolif model}
\usage{
prolif.single.peak(x, mean, sd, summit)
}
\arguments{
\item{x}{x values (log scaled intensity values)}

\item{mean}{the mean of the peak}

\item{sd}{the sd of the peak}

\item{the}{peak height (estimated relative to the other peaks)}
}
\value{
scaled density for peak height
}
\description{
Scaled distribution of a single peak in prolif model
}
