% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{prolif_model}
\alias{prolif_model}
\title{Mixture distribution of Gaussian}
\usage{
prolif_model(means, sd, summits, x)
}
\description{
Gaussian prolif model, it is a mixture distribution scaled by relative
peak heights.
}
