% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{prolif_model_density}
\alias{prolif_model_density}
\title{Density of mixture distribution of Gaussian}
\usage{
prolif_model_density(
  means,
  sd,
  summits,
  x,
  verbose = F,
  log = F,
  opt.env = NULL
)
}
\description{
Density of mixture distribution of Gaussian
}
