% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{prolif_model_nll}
\alias{prolif_model_nll}
\title{Negative log likelihood of mixture distribution of Gaussian}
\usage{
prolif_model_nll(
  means,
  sd,
  summits,
  x,
  verbose = F,
  log = T,
  invert = F,
  opt.env = NULL
)
}
\description{
Negative log likelihood of mixture distribution of Gaussian
}
