% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{prolif_model_wrapper}
\alias{prolif_model_wrapper}
\title{Wrapper that returns a gaussian proliferation model.}
\usage{
prolif_model_wrapper(
  par,
  n.peaks,
  x,
  fixed = NULL,
  type = "prolif_model",
  verbose = F,
  opt.env = NULL,
  invert = F,
  names = NULL,
  log = T
)
}
\description{
Takes a vector of parameters par, number of peaks, and data X
}
