% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{prolif_single_peak}
\alias{prolif_single_peak}
\title{Scaled distribution of a single peak in prolif model}
\usage{
prolif_single_peak(x, mean, sd, summit)
}
\arguments{
\item{x}{x values (log scaled intensity values)}

\item{mean}{the mean of the peak}

\item{sd}{the sd of the peak}

\item{summit}{the peak height (estimated relative to the other peaks)}
}
\value{
scaled density for peak height
}
\description{
Scaled distribution of a single peak in prolif model
}
