% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proliferation_modeling_functions.r
\name{find.enrichment}
\alias{find.enrichment}
\title{Find the relative enrichment over a valley}
\usage{
find.enrichment(x, y, curpeak.mean, est.window)
}
\description{
Find the relative enrichment over a valley
}
