% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proliferation_modeling_functions.r
\name{find.initial.peaks}
\alias{find.initial.peaks}
\title{Estimate initial parameters for proliferation model}
\usage{
find.initial.peaks(
  x,
  y,
  peak.0.lower.bound,
  peak.thresh.enrich = 1,
  peak.thresh.summit = 0.05,
  peak.max = 12,
  plot = F,
  window.scaling.factors = c(1, 1)
)
}
\arguments{
\item{y}{a smoothed or raw count trace}

\item{peak.thresh.enrich}{fold change over valley to call peak in initial estimation (default 1)}

\item{peak.thresh.summit}{minimum height of a peak in percentage of total heights (default 0.05)}

\item{peak.max}{the maximum number of peaks to search for (default 12)}

\item{plot}{should plot be generated}

\item{cur.hist}{histogram object}
}
\value{
data frame with initial peak estimates
}
\description{
Estimate initial parameters for proliferation model
}
