% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_detection.R
\name{find_next_peak}
\alias{find_next_peak}
\title{Find the next peak in a trace}
\usage{
find_next_peak(
  x,
  y,
  prev.peak,
  peak.dist,
  window.scaling.factors = c(0.25, 0.25)
)
}
\description{
Find the mode of the next peak using the previous peak and the average peak
distance. The window can be asymmetrically scaled using the scaling factors.
This is done as the smaller the peaks get, the closer together they tend to be.
}
