% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/peak_detection.R
\name{find.peak.x.approx.mode}
\alias{find.peak.x.approx.mode}
\alias{find_peak_x_approx_mode}
\title{Find the mode of the first left sided peak in the data, given some constraints.}
\usage{
find.peak.x.approx.mode(...)

find_peak_x_approx_mode(
  trace,
  peak.x.thresh.summit,
  peak.x.thresh.enrich,
  peak.x.upper.bound = NULL,
  ...
)
}
\arguments{
\item{...}{additional arguments passed to \code{\link[stats]{density}}.}

\item{trace}{raw FACS intensities (log10 scale)}

\item{peak.x.thresh.summit}{minimum relative height of peak x}

\item{peak.x.thresh.enrich}{minimum enrichment of peak x over valley}

\item{peak.x.upper.bound}{hard upper limit for peak x position on the log10
scale. If the detected mode exceeds this value, NULL is returned.}
}
\value{
the log10 x position of the estimated peak x mode, or NULL if no
  valid peak is found within the specified constraints.
}
\description{
Find the mode of the first left sided peak in the data, given some constraints.
}
