% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto_cocult_functions.r
\name{fit.peaks}
\alias{fit.peaks}
\title{Fit a proliferation model using least squares}
\usage{
fit.peaks(
  trace,
  peak.0.lower.bound,
  bins = 250,
  smoothing.window = 2,
  plot = T,
  plot.main = "Proliferation model",
  opt.peak.pos.dev = NULL,
  opt.trim.left.tail = T,
  mode = "LS",
  ...
)
}
\arguments{
\item{peak.0.lower.bound}{the value on log10 scale where the first valley is}

\item{bins}{number of bins for fitting (default 250)}

\item{smoothing.window}{number of values up and downstream for the NN smoother (default 2)}

\item{plot}{should results be plotted}

\item{plot.main}{title for the plot}

\item{opt.peak.pos.dev}{limits the range where peak positions can be. See @details}

\item{opt.trim.left.tail}{Should values that fall outside the model range bet set to zero @details}

\item{x}{raw FACS intensities}

\item{peak.thresh.enrich}{fold change over valley to call peak in initial estimation (default 1)}

\item{peak.thresh.summit}{minimum height of a peak in percentage of total heights (default 0.05)}

\item{peak.max}{the maximum number of peaks to search for (default 12)}
}
\value{
A data frame with peak statistics
}
\description{
Fit peaks on a CTV or CFSE or other tracking dye trace using a binned approach.
The trace should be raw per event data from the FCS file on the CTV+ population.
}
\details{
As a proliferation model, a guassian mixture distribution is fit.
Initial parameters (number of peaks, peak means, peak summits) are estimated
using a simple approach where first the generation zero peak is estimated based
on 'peak.0.lower.bound' on smoothed data. The bins are smoothed using a
nearest neighbour smoother in the window provided.

Peaks are only called if they exceed peak.thresh.enrich, the relative enrichment
between the valley's either side of the peak and the peak summit. Peaks
must also be adjacent. So if peak 2 does not pass peak.thresh.enrich but peak
3 does, only 2 peaks are fit. This is done to avoid fitting many peaks in the
marginal count range.

Using the gen0 peak the subsequent peak position is estimated, after which
the summit is found and the peak mean updated to the local summit. This is
repeated using each subsequent peak.

After estimating starting values, a gaussian mixture distribution is fit to
the trace (a sum of the PDF of individuals gaussians). The starting values
for the peak standard deviations are estimated from the trace in the bound
of peak0. These values are input into the Levenberg-Marquardt algo to
optimise with respect to the resdiual sum of squares between the model
and the smoothed trace.

opt.peak.pos.dev controls the range peaks are allowed to vary in position
during optimization. Defaults to +- half the estimated peak distance (NULL).
To ignore this, set to Inf. This setting avoids the optimizer putting peaks
in the left tail that sometimes might be present.
}
\examples{

# Simulate proliferation data
y <- 10 ^ rnorm(1000, mean=10, sd=0.05)
y <- c(y/4, y/2, y)

# Fit peaks
peaks <- fit.peaks(y, peak.0.lower.bound=9.8)

# Simulate proliferation data #2
y <- 10 ^ rnorm(1000, mean=10, sd=0.05)
y <- c((y/8)[1:500], y/4, y/2, y[1:500])

peaks <- fit.peaks(y, peak.0.lower.bound=9.9, mode="MLE")
}
