% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proliferation_modeling_functions.r
\name{opt.prolif.model.mle}
\alias{opt.prolif.model.mle}
\title{Optimize a proliferation model using maximum likelihood}
\usage{
opt.prolif.model.mle(
  x,
  starts,
  upper,
  lower,
  fixed,
  peak.stats,
  plot = F,
  opt.env = NULL,
  verbose = F,
  log = T
)
}
\description{
Optimize a proliferation model using maximum likelihood
}
