% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{opt_plot_estimates}
\alias{opt_plot_estimates}
\title{Plot initial parameter estimates}
\usage{
opt_plot_estimates(
  cur.hist,
  y.smth,
  peak.stats,
  peak.0.lower.bound,
  peak.x.model,
  peak.x.mode,
  peak.x.upper.bound
)
}
\description{
Plot initial parameter estimates
}
