% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{opt_plot_final}
\alias{opt_plot_final}
\title{Plot optimization of parameters}
\usage{
opt_plot_final(x.mids, y, y.pred.optim, main = "Optimal fit vs data")
}
\description{
Plot optimization of parameters
}
