% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{opt_prolif_model_ls}
\alias{opt_prolif_model_ls}
\title{Optimize a proliferation model using least squares}
\usage{
opt_prolif_model_ls(
  x,
  y,
  starts,
  upper,
  lower,
  fixed,
  peak.stats,
  plot = F,
  opt.env = NULL
)
}
\description{
Optimize a proliferation model using least squares
}
