% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{opt_prolif_model_mle}
\alias{opt_prolif_model_mle}
\title{Optimize a proliferation model using maximum likelihood}
\usage{
opt_prolif_model_mle(
  x,
  starts,
  upper,
  lower,
  fixed,
  peak.stats,
  plot = F,
  opt.env = NULL,
  verbose = F,
  log = T
)
}
\description{
Optimize a proliferation model using maximum likelihood
}
