% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proliferation_modeling_functions.r
\name{prolif.model}
\alias{prolif.model}
\title{Mixture distribution of Gaussian}
\usage{
prolif.model(means, sd, summits, x)
}
\description{
Gaussian prolif model, it is a mixture distribution scaled by relative
peak heights.
}
