% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proliferation_modeling_functions.r
\name{prolif.model.wrapper}
\alias{prolif.model.wrapper}
\title{Wrapper that returns a gaussian proliferation model.}
\usage{
prolif.model.wrapper(
  par,
  n.peaks,
  x,
  fixed = NULL,
  type = "prolif_model",
  verbose = F,
  opt.env = NULL,
  invert = F,
  names = NULL,
  log = T
)
}
\description{
Takes a vector of paremeters par, number of peaks, and data X
}
