% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{prolif_resid}
\alias{prolif_resid}
\title{Residuals between observed CTV trace and Gaussian prolif model}
\usage{
prolif_resid(par, n.peaks, x, y, fixed = NULL, opt.env = NULL)
}
\description{
Residuals between observed CTV trace and Gaussian prolif model
}
